requirements.txt
python-telegram-bot==21.6
nest-asyncio==1.6.0